requirements.txt
# Core LangChain packages
langchain-core>=1.2.6
langchain-community>=0.4.1
langchain-openai>=1.1.7
langchain-text-splitters>=1.1.0

# Vector database
chromadb>=1.4.0

# PDF processing
pdfplumber>=0.11.9

# OpenAI (dependency for langchain-openai)
openai>=2.14.0

# API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
